refactor: Refactor import paths for work area related types to use module aliases
